Refactor normalization to merge duplicate entities.
Previously, when a resource that was already previously normalized is
encountered during traversal, it would overwrite all of the old values.
This led to issues such as #17, where a reference would overwrite the
actual entity.
With this change, such entities are instead *merged* with the previously
encountered entity, so that only values that were previously empty or
not set will be set on the existing entity. In the case of objects and
arrays, both the old and the new values will be combined, with the
values for the existing entity taking precedence over the new values.